+++ src/custom-code-helpers/string-array/templates/string-array-rotate-function/StringArrayRotateFunctionTemplate.ts
@@ -4,14 +4,11 @@
 export function StringArrayRotateFunctionTemplate (): string {
     return `
         (function (array, comparisonValue) {
-            array[-1] = true;
-
             while (true) {
                 try {
                     const expression = {comparisonExpressionCode};
                                             
                     if (expression === comparisonValue) {
-                        array[-1] = false;
                         break;
                     } else {
                         array['push'](array['shift']());