+++ test/runtime-tests/JavaScriptObfuscatorRuntime.spec.ts
@@ -4,6 +4,8 @@ import { TInputOptions } from '../../src/types/options/TInputOptions';
 
 import { IdentifierNamesGenerator } from '../../src/enums/generators/identifier-names-generators/IdentifierNamesGenerator';
 import { StringArrayEncoding } from '../../src/enums/node-transformers/string-array-transformers/StringArrayEncoding';
+import { StringArrayIndexesType } from '../../src/enums/node-transformers/string-array-transformers/StringArrayIndexesType';
+import { StringArrayWrappersType } from '../../src/enums/node-transformers/string-array-transformers/StringArrayWrappersType';
 
 import { evaluateInWorker } from '../helpers/evaluateInWorker';
 import { readFileAsString } from '../helpers/readFileAsString';
@@ -18,17 +20,19 @@ const getEnvironmentCode = () => `
 
 describe('JavaScriptObfuscator runtime eval', function () {
     const baseOptions: TInputOptions = {
-        compact: false,
+        controlFlowFlattening: true,
+        controlFlowFlatteningThreshold: 1,
         deadCodeInjection: true,
         deadCodeInjectionThreshold: 1,
-        // debugProtection: true,
-        disableConsoleOutput: false,
+        debugProtection: true,
+        disableConsoleOutput: true,
         domainLock: ['obfuscator.io'],
         numbersToExpressions: true,
         simplify: true,
+        renameProperties: true,
         reservedNames: ['generate', 'sha256'],
-        rotateStringArray: false,
-        // selfDefending: true,
+        rotateStringArray: true,
+        selfDefending: true,
         splitStrings: true,
         splitStringsChunkLength: 3,
         stringArray: true,
@@ -37,9 +41,17 @@ describe('JavaScriptObfuscator runtime eval', function () {
             StringArrayEncoding.Base64,
             StringArrayEncoding.Rc4
         ],
+        stringArrayIndexesType: [
+            StringArrayIndexesType.HexadecimalNumber,
+            StringArrayIndexesType.HexadecimalNumericString
+        ],
+        stringArrayIndexShift: true,
         stringArrayWrappersChainedCalls: true,
         stringArrayWrappersCount: 5,
+        stringArrayWrappersParametersMaxCount: 5,
+        stringArrayWrappersType: StringArrayWrappersType.Function,
         stringArrayThreshold: 1,
+        transformObjectKeys: true,
         unicodeEscapeSequence: true
     };